+++ src/core/opcua-iiot-core.ts
@@ -538,10 +538,10 @@ export function convertDataValueByDataType(value: any, dataType: DataTypeInput):
         break;
       default:
         logger.internalDebugLog('convertDataValue unused DataType: ' + dataType)
-        if (!isNotDefined(value)) {
-          convertedValue = value
-        } else {
+        if(_.isUndefined(value)) {
           convertedValue = null
+        } else {
+          convertedValue = value
         }
         break
     }
@@ -978,7 +978,7 @@ export function registerToConnector(node: TodoTypeAny, statusCallback: (status: 
   node.connector.on('after_reconnection', () => {
     setNodeOPCUARestart(node.connector, OPCUAClient.create((node.connector as TodoTypeAny).iiot.opcuaClient), statusCall) // TODO: investigate one args v two
   })
-  setNodeInitalState(node.connector?.iiot?.stateService.state.value, node, statusCall) //Todo: switch to xstate/fsm from stately
+  setNodeInitalState(node.connector?.iiot?.stateService.state.value, node, statusCall)
 }
 
 export function deregisterToConnector(node: NodeWithConnector, done: () => void) {