+++ test/e2e/opcua-iiot-connector-e2e.test.js
@@ -342,109 +342,123 @@ describe('OPC UA Connector node e2e Testing', function () {
           // TODO: nodeClass is now int and not string anymore
           expect(msg.payload.browserResults).toMatchObject([
             {
-              'nodeId': 'ns=1;s=Pressure',
-              'browseName': '1:Pressure',
-              'displayName': 'locale=en-US text=Pressure',
-              'nodeClass': '2',
-              'datatypeName': 'ns=0;i=63'
+              "nodeId": "ns=1;s=Pressure",
+              "browseName": "1:Pressure",
+              "displayName": "locale=en-US text=P4NR Pressure",
+              "nodeClass": "2",
+              "datatypeName": "ns=0;i=63"
             },
             {
-              'nodeId': 'ns=1;s=Matrix',
-              'browseName': '1:Matrix',
-              'displayName': 'locale=en-US text=Matrix',
-              'nodeClass': '2',
-              'datatypeName': 'ns=0;i=63'
+              "nodeId": "ns=1;s=Matrix",
+              "browseName": "1:Matrix",
+              "displayName": "locale=en-US text=P4NR Matrix",
+              "nodeClass": "2",
+              "datatypeName": "ns=0;i=63"
             },
             {
-              'nodeId': 'ns=1;s=Position',
-              'browseName': '1:Position',
-              'displayName': 'locale=en-US text=Position',
-              'nodeClass': '2',
-              'datatypeName': 'ns=0;i=63'
+              "nodeId": "ns=1;s=Position",
+              "browseName": "1:Position",
+              "displayName": "locale=en-US text=P4NR Position",
+              "nodeClass": "2",
+              "datatypeName": "ns=0;i=63"
             },
             {
-              'nodeId': 'ns=1;s=PumpSpeed',
-              'browseName': '1:PumpSpeed',
-              'displayName': 'locale=en-US text=Pump Speed',
-              'nodeClass': '2',
-              'datatypeName': 'ns=0;i=63'
+              "nodeId": "ns=1;s=PumpSpeed",
+              "browseName": "1:PumpSpeed",
+              "displayName": "locale=en-US text=P4NR Pump Speed",
+              "nodeClass": "2",
+              "datatypeName": "ns=0;i=63"
             },
             {
-              'nodeId': 'ns=1;s=SomeDate',
-              'browseName': '1:SomeDate',
-              'displayName': 'locale=en-US text=Some Date',
-              'nodeClass': '2',
-              'datatypeName': 'ns=0;i=63'
+              "nodeId": "ns=1;s=SomeDate",
+              "browseName": "1:SomeDate",
+              "displayName": "locale=en-US text=P4NR Some Date",
+              "nodeClass": "2",
+              "datatypeName": "ns=0;i=63"
             },
             {
-              'nodeId': 'ns=1;s=MultiLanguageText',
-              'browseName': '1:MultiLanguageText',
-              'displayName': 'locale=en-US text=Multi Language Text',
-              'nodeClass': '2',
-              'datatypeName': 'ns=0;i=63'
+              "nodeId": "ns=1;s=MultiLanguageText",
+              "browseName": "1:MultiLanguageText",
+              "displayName": "locale=en-US text=P4NR Multi Language Text",
+              "nodeClass": "2",
+              "datatypeName": "ns=0;i=63"
             },
             {
-              'nodeId': 'ns=1;s=FanSpeed',
-              'browseName': '1:FanSpeed',
-              'displayName': 'locale=en-US text=Fan Speed',
-              'nodeClass': '2',
-              'datatypeName': 'ns=0;i=63'
+              "nodeId": "ns=1;s=FanSpeed",
+              "browseName": "1:FanSpeed",
+              "displayName": "locale=en-US text=P4NR Fan Speed",
+              "nodeClass": "2",
+              "datatypeName": "ns=0;i=63"
             },
             {
-              'nodeId': 'ns=1;s=TemperatureAnalogItem',
-              'browseName': '1:TemperatureAnalogItem',
-              'displayName': 'locale=en-US text=Temperature',
-              'nodeClass': '2',
-              'datatypeName': 'ns=0;i=2368'
+              "nodeId": "ns=1;i=16479",
+              "browseName": "1:MyVariable1",
+              "displayName": "locale=null text=PLUS for Node-RED Double Variable",
+              "nodeClass": "2",
+              "datatypeName": "ns=0;i=63"
             },
             {
-              'nodeId': 'ns=1;i=16479',
-              'browseName': '1:MyVariable1',
-              'displayName': 'locale=null text=My Variable 1',
-              'nodeClass': '2',
-              'datatypeName': 'ns=0;i=63'
+              "nodeId": "ns=1;b=1020FFAA",
+              "browseName": "1:MyVariable2",
+              "displayName": "locale=null text=PLUS for Node-RED Double Binary Variable",
+              "nodeClass": "2",
+              "datatypeName": "ns=0;i=63"
             },
             {
-              'nodeId': 'ns=1;b=1020FFAA',
-              'browseName': '1:MyVariable2',
-              'displayName': 'locale=null text=My Variable 2',
-              'nodeClass': '2',
-              'datatypeName': 'ns=0;i=63'
+              "nodeId": "ns=1;s=TestReadWrite",
+              "browseName": "1:TestReadWrite",
+              "displayName": "locale=null text=PLUS for Node-RED Test Read and Write",
+              "nodeClass": "2",
+              "datatypeName": "ns=0;i=63"
             },
             {
-              'nodeId': 'ns=1;s=TestReadWrite',
-              'browseName': '1:TestReadWrite',
-              'displayName': 'locale=null text=Test Read and Write',
-              'nodeClass': '2',
-              'datatypeName': 'ns=0;i=63'
+              "nodeId": "ns=1;s=free_memory",
+              "browseName": "1:FreeMemory",
+              "displayName": "locale=en-US text=P4NR Free Memory",
+              "nodeClass": "2",
+              "datatypeName": "ns=0;i=63"
             },
             {
-              'nodeId': 'ns=1;s=free_memory',
-              'browseName': '1:FreeMemory',
-              'displayName': 'locale=en-US text=Free Memory',
-              'nodeClass': '2',
-              'datatypeName': 'ns=0;i=63'
+              "nodeId": "ns=1;s=Counter",
+              "browseName": "1:Counter",
+              "displayName": "locale=en-US text=P4NR Counter",
+              "nodeClass": "2",
+              "datatypeName": "ns=0;i=63"
             },
             {
-              'nodeId': 'ns=1;s=Counter',
-              'browseName': '1:Counter',
-              'displayName': 'locale=en-US text=Counter',
-              'nodeClass': '2',
-              'datatypeName': 'ns=0;i=63'
+              "nodeId": "ns=1;s=FullCounter",
+              "browseName": "1:FullCounter",
+              "displayName": "locale=en-US text=P4NR Full-Counter",
+              "nodeClass": "2",
+              "datatypeName": "ns=0;i=63"
             },
             {
-              'nodeId': 'ns=1;s=FullCounter',
-              'browseName': '1:FullCounter',
-              'displayName': 'locale=en-US text=Full-Counter',
-              'nodeClass': '2',
-              'datatypeName': 'ns=0;i=63'
+              "nodeId": "ns=1;i=1143",
+              "browseName": "1:FullCounterLimitCondition",
+              "displayName": "locale=en-US text=Full Counter Limit Condition",
+              "nodeClass": "1",
+              "datatypeName": "ns=0;i=9906"
             },
             {
-              'nodeId': 'ns=1;i=12345',
-              'browseName': '1:Bark',
-              'displayName': 'locale=null text=Bark',
-              'nodeClass': '4',
-              'datatypeName': 'ns=0;i=0'
+              "nodeId": "ns=1;i=12345",
+              "browseName": "1:Bark",
+              "displayName": "locale=null text=Bark",
+              "nodeClass": "4",
+              "datatypeName": "ns=0;i=0"
+            },
+            {
+              "nodeId": "ns=1;s=TemperatureAnalogItem",
+              "browseName": "1:TemperatureAnalogItem",
+              "displayName": "locale=en-US text=P4NR Temperature",
+              "nodeClass": "2",
+              "datatypeName": "ns=0;i=2368"
+            },
+            {
+              "nodeId": "ns=1;i=1201",
+              "browseName": "1:TemperatureLimitCondition",
+              "displayName": "locale=en-US text=Temperature Limit Condition",
+              "nodeClass": "1",
+              "datatypeName": "ns=0;i=9341"
             }
           ])
           done()